refactor(codegen): extrai passes/args/* de func.rs
Move 3 passes de reescrita de argumentos (763 linhas) para
crates/rts-codegen/src/codegen/lower/passes/args/:
- default_args.rs (264 LOC) — preenche defaults omitidos no callsite
- rest_args.rs (245 LOC) — empacota ...rest em array literal
- spread_args.rs (254 LOC) — expande ...arr inline em args

func.rs: 8175 -> 7413 linhas. Suite verde.
