+++ crates/rts-codegen/src/codegen/lower/func.rs
@@ -26,6 +26,9 @@ use super::analysis::captures::{
 };
 use super::analysis::module_globals::collect_module_globals;
 use super::analysis::types::sanitize_symbol;
+use super::passes::args::default_args::expand_default_args;
+use super::passes::args::rest_args::expand_rest_args;
+use super::passes::args::spread_args::expand_spread_args;
 use super::ctx::{ClassMeta, FnCtx, GlobalVar, UserFnAbi, ValTy};
 use super::statements::lower_stmt;
 
@@ -3948,771 +3951,6 @@ fn push_array_rest_decl(
     out.push(Statement::Raw(
         RawStmt::new("<destruct-rest-loop>".to_string(), Span::default()).with_stmt(for_stmt),
     ));
-}
-
-fn expand_default_args(program: &mut Program) {
-    use std::collections::HashMap;
-
-    // Mapa: nome → params (defaults inclusos). Para métodos: mesmo nome
-    // pode aparecer em múltiplas classes — guardamos em outro mapa
-    // indexado por (class, method).
-    let mut fn_defaults: HashMap<String, Vec<Option<Box<Expr>>>> = HashMap::new();
-    let mut method_defaults: HashMap<(String, String), Vec<Option<Box<Expr>>>> = HashMap::new();
-
-    for item in &program.items {
-        match item {
-            Item::Function(f) => {
-                if f.parameters.iter().any(|p| p.default.is_some()) {
-                    let defaults: Vec<Option<Box<Expr>>> =
-                        f.parameters.iter().map(|p| p.default.clone()).collect();
-                    fn_defaults.insert(f.name.clone(), defaults);
-                }
-            }
-            Item::Class(c) => {
-                for m in &c.members {
-                    if let ClassMember::Method(method) = m {
-                        if method.parameters.iter().any(|p| p.default.is_some()) {
-                            let defaults: Vec<Option<Box<Expr>>> = method
-                                .parameters
-                                .iter()
-                                .map(|p| p.default.clone())
-                                .collect();
-                            method_defaults.insert((c.name.clone(), method.name.clone()), defaults);
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    if fn_defaults.is_empty() && method_defaults.is_empty() {
-        return;
-    }
-
-    // Reescreve callsites.
-    for item in program.items.iter_mut() {
-        match item {
-            Item::Function(f) => {
-                for s in f.body.iter_mut() {
-                    let Statement::Raw(raw) = s;
-                    if let Some(stmt) = raw.stmt.as_mut() {
-                        expand_in_stmt(stmt, &fn_defaults, &method_defaults);
-                    }
-                }
-            }
-            Item::Class(c) => {
-                for m in c.members.iter_mut() {
-                    match m {
-                        ClassMember::Constructor(ctor) => {
-                            for s in ctor.body.iter_mut() {
-                                let Statement::Raw(raw) = s;
-                                if let Some(stmt) = raw.stmt.as_mut() {
-                                    expand_in_stmt(stmt, &fn_defaults, &method_defaults);
-                                }
-                            }
-                        }
-                        ClassMember::Method(method) => {
-                            for s in method.body.iter_mut() {
-                                let Statement::Raw(raw) = s;
-                                if let Some(stmt) = raw.stmt.as_mut() {
-                                    expand_in_stmt(stmt, &fn_defaults, &method_defaults);
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            Item::Statement(Statement::Raw(raw)) => {
-                if let Some(stmt) = raw.stmt.as_mut() {
-                    expand_in_stmt(stmt, &fn_defaults, &method_defaults);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-fn expand_in_stmt(
-    stmt: &mut Stmt,
-    fn_defaults: &std::collections::HashMap<String, Vec<Option<Box<Expr>>>>,
-    method_defaults: &std::collections::HashMap<(String, String), Vec<Option<Box<Expr>>>>,
-) {
-    use swc_ecma_ast::Stmt::*;
-    match stmt {
-        Expr(e) => expand_in_expr(&mut e.expr, fn_defaults, method_defaults),
-        Return(r) => {
-            if let Some(a) = r.arg.as_deref_mut() {
-                expand_in_expr(a, fn_defaults, method_defaults);
-            }
-        }
-        If(i) => {
-            expand_in_expr(&mut i.test, fn_defaults, method_defaults);
-            expand_in_stmt(&mut i.cons, fn_defaults, method_defaults);
-            if let Some(alt) = i.alt.as_deref_mut() {
-                expand_in_stmt(alt, fn_defaults, method_defaults);
-            }
-        }
-        Block(b) => {
-            for s in &mut b.stmts {
-                expand_in_stmt(s, fn_defaults, method_defaults);
-            }
-        }
-        While(w) => {
-            expand_in_expr(&mut w.test, fn_defaults, method_defaults);
-            expand_in_stmt(&mut w.body, fn_defaults, method_defaults);
-        }
-        DoWhile(w) => {
-            expand_in_expr(&mut w.test, fn_defaults, method_defaults);
-            expand_in_stmt(&mut w.body, fn_defaults, method_defaults);
-        }
-        For(f) => {
-            if let Some(init) = f.init.as_mut() {
-                if let swc_ecma_ast::VarDeclOrExpr::VarDecl(vd) = init {
-                    for d in &mut vd.decls {
-                        if let Some(e) = d.init.as_deref_mut() {
-                            expand_in_expr(e, fn_defaults, method_defaults);
-                        }
-                    }
-                }
-            }
-            if let Some(t) = f.test.as_deref_mut() {
-                expand_in_expr(t, fn_defaults, method_defaults);
-            }
-            if let Some(u) = f.update.as_deref_mut() {
-                expand_in_expr(u, fn_defaults, method_defaults);
-            }
-            expand_in_stmt(&mut f.body, fn_defaults, method_defaults);
-        }
-        ForOf(f) => {
-            expand_in_expr(&mut f.right, fn_defaults, method_defaults);
-            expand_in_stmt(&mut f.body, fn_defaults, method_defaults);
-        }
-        Decl(swc_ecma_ast::Decl::Var(v)) => {
-            for d in &mut v.decls {
-                if let Some(e) = d.init.as_deref_mut() {
-                    expand_in_expr(e, fn_defaults, method_defaults);
-                }
-            }
-        }
-        Try(t) => {
-            for s in &mut t.block.stmts {
-                expand_in_stmt(s, fn_defaults, method_defaults);
-            }
-            if let Some(h) = t.handler.as_mut() {
-                for s in &mut h.body.stmts {
-                    expand_in_stmt(s, fn_defaults, method_defaults);
-                }
-            }
-            if let Some(f) = t.finalizer.as_mut() {
-                for s in &mut f.stmts {
-                    expand_in_stmt(s, fn_defaults, method_defaults);
-                }
-            }
-        }
-        _ => {}
-    }
-}
-
-fn expand_in_expr(
-    expr: &mut Expr,
-    fn_defaults: &std::collections::HashMap<String, Vec<Option<Box<Expr>>>>,
-    method_defaults: &std::collections::HashMap<(String, String), Vec<Option<Box<Expr>>>>,
-) {
-    // Recurse primeiro para que callsites internos também sejam expandidos.
-    match expr {
-        Expr::Call(call) => {
-            // Recurse em args primeiro (call aninhado).
-            for a in call.args.iter_mut() {
-                expand_in_expr(&mut a.expr, fn_defaults, method_defaults);
-            }
-            if let Callee::Expr(callee_expr) = &mut call.callee {
-                expand_in_expr(callee_expr, fn_defaults, method_defaults);
-            }
-            // Detecta callee:
-            //   - Ident("f") → fn_defaults["f"]
-            //   - Member(this, "m") em método de classe → não temos
-            //     contexto da classe aqui, então skip; será tratado em
-            //     dispatch virtual posterior.
-            //   - Member(obj_local, "m") onde obj_local é Ident — skip por
-            //     mesmo motivo.
-            //
-            // Cobertura: defaults de fns top-level. Defaults em métodos
-            // ficam parcialmente cobertos no codegen futuro (não nesse
-            // pass).
-            let fn_name = if let Callee::Expr(ce) = &call.callee {
-                if let Expr::Ident(id) = ce.as_ref() {
-                    Some(id.sym.to_string())
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-            if let Some(name) = fn_name {
-                if let Some(defaults) = fn_defaults.get(&name) {
-                    let provided = call.args.len();
-                    let total = defaults.len();
-                    if provided < total {
-                        for i in provided..total {
-                            if let Some(def) = &defaults[i] {
-                                let mut def_clone = (**def).clone();
-                                expand_in_expr(&mut def_clone, fn_defaults, method_defaults);
-                                call.args.push(swc_ecma_ast::ExprOrSpread {
-                                    spread: None,
-                                    expr: Box::new(def_clone),
-                                });
-                            } else {
-                                // Param sem default — TS exige que callsite
-                                // proveja, codegen vai dar erro mais claro.
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Expr::Member(m) => expand_in_expr(&mut m.obj, fn_defaults, method_defaults),
-        Expr::Bin(b) => {
-            expand_in_expr(&mut b.left, fn_defaults, method_defaults);
-            expand_in_expr(&mut b.right, fn_defaults, method_defaults);
-        }
-        Expr::Unary(u) => expand_in_expr(&mut u.arg, fn_defaults, method_defaults),
-        Expr::Update(u) => expand_in_expr(&mut u.arg, fn_defaults, method_defaults),
-        Expr::Assign(a) => {
-            if let swc_ecma_ast::AssignTarget::Simple(swc_ecma_ast::SimpleAssignTarget::Member(m)) =
-                &mut a.left
-            {
-                expand_in_expr(&mut m.obj, fn_defaults, method_defaults);
-            }
-            expand_in_expr(&mut a.right, fn_defaults, method_defaults);
-        }
-        Expr::New(n) => {
-            if let Some(args) = n.args.as_mut() {
-                for a in args {
-                    expand_in_expr(&mut a.expr, fn_defaults, method_defaults);
-                }
-            }
-        }
-        Expr::Cond(c) => {
-            expand_in_expr(&mut c.test, fn_defaults, method_defaults);
-            expand_in_expr(&mut c.cons, fn_defaults, method_defaults);
-            expand_in_expr(&mut c.alt, fn_defaults, method_defaults);
-        }
-        Expr::Paren(p) => expand_in_expr(&mut p.expr, fn_defaults, method_defaults),
-        Expr::Tpl(t) => {
-            for e in &mut t.exprs {
-                expand_in_expr(e, fn_defaults, method_defaults);
-            }
-        }
-        Expr::Array(a) => {
-            for el in a.elems.iter_mut().flatten() {
-                expand_in_expr(&mut el.expr, fn_defaults, method_defaults);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// Empacota argumentos variádicos `...rest` num array literal no callsite.
-///
-/// Para uma fn user com último param marcado \`variadic\` (sintaxe
-/// \`...rest\`), todos os args do callsite a partir da posição desse param
-/// são coletados num \`Expr::Array\` e passados como único arg na posição
-/// do rest. Codegen do callee vê \`rest\` como Handle de array normal —
-/// pode iterar via \`for...of\`.
-fn expand_rest_args(program: &mut Program) {
-    use std::collections::HashMap;
-
-    // Mapa: nome → índice do parâmetro variadic (último). Apenas
-    // funções top-level e métodos.
-    let mut fn_rest: HashMap<String, usize> = HashMap::new();
-    let mut method_rest: HashMap<(String, String), usize> = HashMap::new();
-
-    for item in &program.items {
-        match item {
-            Item::Function(f) => {
-                if let Some(idx) = f.parameters.iter().position(|p| p.variadic) {
-                    fn_rest.insert(f.name.clone(), idx);
-                }
-            }
-            Item::Class(c) => {
-                for m in &c.members {
-                    if let ClassMember::Method(method) = m {
-                        if let Some(idx) = method.parameters.iter().position(|p| p.variadic) {
-                            method_rest.insert((c.name.clone(), method.name.clone()), idx);
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    if fn_rest.is_empty() && method_rest.is_empty() {
-        return;
-    }
-
-    // Reescrita.
-    for item in program.items.iter_mut() {
-        match item {
-            Item::Function(f) => {
-                for s in f.body.iter_mut() {
-                    let Statement::Raw(raw) = s;
-                    if let Some(stmt) = raw.stmt.as_mut() {
-                        rest_in_stmt(stmt, &fn_rest, &method_rest);
-                    }
-                }
-            }
-            Item::Class(c) => {
-                for m in c.members.iter_mut() {
-                    match m {
-                        ClassMember::Constructor(ctor) => {
-                            for s in ctor.body.iter_mut() {
-                                let Statement::Raw(raw) = s;
-                                if let Some(stmt) = raw.stmt.as_mut() {
-                                    rest_in_stmt(stmt, &fn_rest, &method_rest);
-                                }
-                            }
-                        }
-                        ClassMember::Method(method) => {
-                            for s in method.body.iter_mut() {
-                                let Statement::Raw(raw) = s;
-                                if let Some(stmt) = raw.stmt.as_mut() {
-                                    rest_in_stmt(stmt, &fn_rest, &method_rest);
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            Item::Statement(Statement::Raw(raw)) => {
-                if let Some(stmt) = raw.stmt.as_mut() {
-                    rest_in_stmt(stmt, &fn_rest, &method_rest);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-fn rest_in_stmt(
-    stmt: &mut Stmt,
-    fn_rest: &std::collections::HashMap<String, usize>,
-    method_rest: &std::collections::HashMap<(String, String), usize>,
-) {
-    use swc_ecma_ast::Stmt::*;
-    match stmt {
-        Expr(e) => rest_in_expr(&mut e.expr, fn_rest, method_rest),
-        Return(r) => {
-            if let Some(a) = r.arg.as_deref_mut() {
-                rest_in_expr(a, fn_rest, method_rest);
-            }
-        }
-        If(i) => {
-            rest_in_expr(&mut i.test, fn_rest, method_rest);
-            rest_in_stmt(&mut i.cons, fn_rest, method_rest);
-            if let Some(alt) = i.alt.as_deref_mut() {
-                rest_in_stmt(alt, fn_rest, method_rest);
-            }
-        }
-        Block(b) => {
-            for s in &mut b.stmts {
-                rest_in_stmt(s, fn_rest, method_rest);
-            }
-        }
-        While(w) => {
-            rest_in_expr(&mut w.test, fn_rest, method_rest);
-            rest_in_stmt(&mut w.body, fn_rest, method_rest);
-        }
-        DoWhile(w) => {
-            rest_in_expr(&mut w.test, fn_rest, method_rest);
-            rest_in_stmt(&mut w.body, fn_rest, method_rest);
-        }
-        For(f) => {
-            if let Some(init) = f.init.as_mut() {
-                if let swc_ecma_ast::VarDeclOrExpr::VarDecl(vd) = init {
-                    for d in &mut vd.decls {
-                        if let Some(e) = d.init.as_deref_mut() {
-                            rest_in_expr(e, fn_rest, method_rest);
-                        }
-                    }
-                }
-            }
-            if let Some(t) = f.test.as_deref_mut() {
-                rest_in_expr(t, fn_rest, method_rest);
-            }
-            if let Some(u) = f.update.as_deref_mut() {
-                rest_in_expr(u, fn_rest, method_rest);
-            }
-            rest_in_stmt(&mut f.body, fn_rest, method_rest);
-        }
-        ForOf(f) => {
-            rest_in_expr(&mut f.right, fn_rest, method_rest);
-            rest_in_stmt(&mut f.body, fn_rest, method_rest);
-        }
-        Decl(swc_ecma_ast::Decl::Var(v)) => {
-            for d in &mut v.decls {
-                if let Some(e) = d.init.as_deref_mut() {
-                    rest_in_expr(e, fn_rest, method_rest);
-                }
-            }
-        }
-        _ => {}
-    }
-}
-
-fn rest_in_expr(
-    expr: &mut Expr,
-    fn_rest: &std::collections::HashMap<String, usize>,
-    method_rest: &std::collections::HashMap<(String, String), usize>,
-) {
-    match expr {
-        Expr::Call(call) => {
-            // Recurse args/callee primeiro.
-            for a in call.args.iter_mut() {
-                rest_in_expr(&mut a.expr, fn_rest, method_rest);
-            }
-            if let Callee::Expr(callee_expr) = &mut call.callee {
-                rest_in_expr(callee_expr, fn_rest, method_rest);
-            }
-            // Detecta callee Ident → fn_rest.
-            let fn_name = if let Callee::Expr(ce) = &call.callee {
-                if let Expr::Ident(id) = ce.as_ref() {
-                    Some(id.sym.to_string())
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-            if let Some(name) = fn_name {
-                if let Some(&rest_idx) = fn_rest.get(&name) {
-                    pack_rest_args(&mut call.args, rest_idx);
-                }
-            }
-        }
-        Expr::Member(m) => rest_in_expr(&mut m.obj, fn_rest, method_rest),
-        Expr::Bin(b) => {
-            rest_in_expr(&mut b.left, fn_rest, method_rest);
-            rest_in_expr(&mut b.right, fn_rest, method_rest);
-        }
-        Expr::Unary(u) => rest_in_expr(&mut u.arg, fn_rest, method_rest),
-        Expr::Update(u) => rest_in_expr(&mut u.arg, fn_rest, method_rest),
-        Expr::Assign(a) => {
-            if let swc_ecma_ast::AssignTarget::Simple(swc_ecma_ast::SimpleAssignTarget::Member(m)) =
-                &mut a.left
-            {
-                rest_in_expr(&mut m.obj, fn_rest, method_rest);
-            }
-            rest_in_expr(&mut a.right, fn_rest, method_rest);
-        }
-        Expr::New(n) => {
-            if let Some(args) = n.args.as_mut() {
-                for a in args {
-                    rest_in_expr(&mut a.expr, fn_rest, method_rest);
-                }
-            }
-        }
-        Expr::Cond(c) => {
-            rest_in_expr(&mut c.test, fn_rest, method_rest);
-            rest_in_expr(&mut c.cons, fn_rest, method_rest);
-            rest_in_expr(&mut c.alt, fn_rest, method_rest);
-        }
-        Expr::Paren(p) => rest_in_expr(&mut p.expr, fn_rest, method_rest),
-        Expr::Tpl(t) => {
-            for e in &mut t.exprs {
-                rest_in_expr(e, fn_rest, method_rest);
-            }
-        }
-        Expr::Array(a) => {
-            for el in a.elems.iter_mut().flatten() {
-                rest_in_expr(&mut el.expr, fn_rest, method_rest);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// Substitui args[rest_idx..] por um Expr::Array contendo esses elementos.
-fn pack_rest_args(args: &mut Vec<swc_ecma_ast::ExprOrSpread>, rest_idx: usize) {
-    if args.len() <= rest_idx {
-        // Caller não passou nenhum arg variadic → empacota array vazio.
-        let empty = Expr::Array(swc_ecma_ast::ArrayLit {
-            span: Default::default(),
-            elems: Vec::new(),
-        });
-        args.push(swc_ecma_ast::ExprOrSpread {
-            spread: None,
-            expr: Box::new(empty),
-        });
-        return;
-    }
-    let extra: Vec<Option<swc_ecma_ast::ExprOrSpread>> = args.drain(rest_idx..).map(Some).collect();
-    let arr = Expr::Array(swc_ecma_ast::ArrayLit {
-        span: Default::default(),
-        elems: extra,
-    });
-    args.push(swc_ecma_ast::ExprOrSpread {
-        spread: None,
-        expr: Box::new(arr),
-    });
-}
-
-/// Expande argumentos com spread literal em callsites: \`fn(...[1,2,3])\`
-/// vira \`fn(1, 2, 3)\` em compile-time. Trabalha sobre toda a árvore.
-///
-/// Cobertura: spread de \`Expr::Array\` literal inline + spread de
-/// \`Ident\` cujo binding (\`const X = [lit, lit, ...]\`) eh resolvivel
-/// estaticamente em compile-time.
-fn expand_spread_args(program: &mut Program) {
-    // Coleta bindings `const X = [...]` em top-level. Map name -> elements.
-    // Replicamos para escopo de fn quando possivel (mantemos apenas scope
-    // top-level por enquanto — bindings locais dentro de fn nao sao
-    // resolvidos aqui, mas podem ser estendidos).
-    let mut const_arrays: std::collections::HashMap<String, swc_ecma_ast::ArrayLit> =
-        std::collections::HashMap::new();
-    for item in &program.items {
-        if let Item::Statement(Statement::Raw(raw)) = item {
-            if let Some(Stmt::Decl(swc_ecma_ast::Decl::Var(vd))) = raw.stmt.as_ref() {
-                if matches!(vd.kind, swc_ecma_ast::VarDeclKind::Const) {
-                    for d in &vd.decls {
-                        if let (Pat::Ident(id), Some(init)) = (&d.name, d.init.as_deref()) {
-                            if let Expr::Array(arr) = init {
-                                // Apenas arrays sem spread interno (1 nivel raso)
-                                // pra evitar reentrancy complexa.
-                                let safe = arr
-                                    .elems
-                                    .iter()
-                                    .all(|e| e.as_ref().map(|x| x.spread.is_none()).unwrap_or(true));
-                                if safe {
-                                    const_arrays
-                                        .insert(id.id.sym.to_string(), arr.clone());
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    for item in program.items.iter_mut() {
-        match item {
-            Item::Function(f) => {
-                for s in f.body.iter_mut() {
-                    let Statement::Raw(raw) = s;
-                    if let Some(stmt) = raw.stmt.as_mut() {
-                        spread_in_stmt(stmt, &const_arrays);
-                    }
-                }
-            }
-            Item::Class(c) => {
-                for m in c.members.iter_mut() {
-                    match m {
-                        ClassMember::Constructor(ctor) => {
-                            for s in ctor.body.iter_mut() {
-                                let Statement::Raw(raw) = s;
-                                if let Some(stmt) = raw.stmt.as_mut() {
-                                    spread_in_stmt(stmt, &const_arrays);
-                                }
-                            }
-                        }
-                        ClassMember::Method(method) => {
-                            for s in method.body.iter_mut() {
-                                let Statement::Raw(raw) = s;
-                                if let Some(stmt) = raw.stmt.as_mut() {
-                                    spread_in_stmt(stmt, &const_arrays);
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            Item::Statement(Statement::Raw(raw)) => {
-                if let Some(stmt) = raw.stmt.as_mut() {
-                    spread_in_stmt(stmt, &const_arrays);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-fn spread_in_stmt(
-    stmt: &mut Stmt,
-    consts: &std::collections::HashMap<String, swc_ecma_ast::ArrayLit>,
-) {
-    use swc_ecma_ast::Stmt::*;
-    match stmt {
-        Expr(e) => spread_in_expr(&mut e.expr, consts),
-        Return(r) => {
-            if let Some(a) = r.arg.as_deref_mut() {
-                spread_in_expr(a, consts);
-            }
-        }
-        If(i) => {
-            spread_in_expr(&mut i.test, consts);
-            spread_in_stmt(&mut i.cons, consts);
-            if let Some(alt) = i.alt.as_deref_mut() {
-                spread_in_stmt(alt, consts);
-            }
-        }
-        Block(b) => {
-            for s in &mut b.stmts {
-                spread_in_stmt(s, consts);
-            }
-        }
-        While(w) => {
-            spread_in_expr(&mut w.test, consts);
-            spread_in_stmt(&mut w.body, consts);
-        }
-        DoWhile(w) => {
-            spread_in_expr(&mut w.test, consts);
-            spread_in_stmt(&mut w.body, consts);
-        }
-        For(f) => {
-            if let Some(init) = f.init.as_mut() {
-                if let swc_ecma_ast::VarDeclOrExpr::VarDecl(vd) = init {
-                    for d in &mut vd.decls {
-                        if let Some(e) = d.init.as_deref_mut() {
-                            spread_in_expr(e, consts);
-                        }
-                    }
-                }
-            }
-            if let Some(t) = f.test.as_deref_mut() {
-                spread_in_expr(t, consts);
-            }
-            if let Some(u) = f.update.as_deref_mut() {
-                spread_in_expr(u, consts);
-            }
-            spread_in_stmt(&mut f.body, consts);
-        }
-        ForOf(f) => {
-            spread_in_expr(&mut f.right, consts);
-            spread_in_stmt(&mut f.body, consts);
-        }
-        Decl(swc_ecma_ast::Decl::Var(v)) => {
-            for d in &mut v.decls {
-                if let Some(e) = d.init.as_deref_mut() {
-                    spread_in_expr(e, consts);
-                }
-            }
-        }
-        _ => {}
-    }
-}
-
-fn spread_in_expr(
-    expr: &mut Expr,
-    consts: &std::collections::HashMap<String, swc_ecma_ast::ArrayLit>,
-) {
-    match expr {
-        Expr::Call(call) => {
-            // Recurse primeiro nos args.
-            for a in call.args.iter_mut() {
-                spread_in_expr(&mut a.expr, consts);
-            }
-            if let Callee::Expr(e) = &mut call.callee {
-                spread_in_expr(e, consts);
-            }
-            expand_spread_in_args(&mut call.args, consts);
-        }
-        Expr::New(n) => {
-            if let Some(args) = n.args.as_mut() {
-                for a in args.iter_mut() {
-                    spread_in_expr(&mut a.expr, consts);
-                }
-                expand_spread_in_args(args, consts);
-            }
-        }
-        Expr::Member(m) => spread_in_expr(&mut m.obj, consts),
-        Expr::Bin(b) => {
-            spread_in_expr(&mut b.left, consts);
-            spread_in_expr(&mut b.right, consts);
-        }
-        Expr::Unary(u) => spread_in_expr(&mut u.arg, consts),
-        Expr::Update(u) => spread_in_expr(&mut u.arg, consts),
-        Expr::Assign(a) => {
-            if let swc_ecma_ast::AssignTarget::Simple(swc_ecma_ast::SimpleAssignTarget::Member(m)) =
-                &mut a.left
-            {
-                spread_in_expr(&mut m.obj, consts);
-            }
-            spread_in_expr(&mut a.right, consts);
-        }
-        Expr::Cond(c) => {
-            spread_in_expr(&mut c.test, consts);
-            spread_in_expr(&mut c.cons, consts);
-            spread_in_expr(&mut c.alt, consts);
-        }
-        Expr::Paren(p) => spread_in_expr(&mut p.expr, consts),
-        Expr::Tpl(t) => {
-            for e in &mut t.exprs {
-                spread_in_expr(e, consts);
-            }
-        }
-        Expr::Array(a) => {
-            for el in a.elems.iter_mut().flatten() {
-                spread_in_expr(&mut el.expr, consts);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// Substitui args com spread de array literal pelos elementos do array.
-/// `f(a, ...[b, c], d)` → `f(a, b, c, d)`.
-/// Tambem resolve `f(...x)` quando `x` eh ident referenciando const top-level
-/// inicializado com array literal estatico.
-fn expand_spread_in_args(
-    args: &mut Vec<swc_ecma_ast::ExprOrSpread>,
-    consts: &std::collections::HashMap<String, swc_ecma_ast::ArrayLit>,
-) {
-    if !args.iter().any(|a| a.spread.is_some()) {
-        return;
-    }
-    let original = std::mem::take(args);
-    for arg in original {
-        if arg.spread.is_some() {
-            // Spread de literal array: expande inline.
-            let arr_opt: Option<&swc_ecma_ast::ArrayLit> = match arg.expr.as_ref() {
-                Expr::Array(arr) => Some(arr),
-                Expr::Ident(id) => consts.get(id.sym.as_ref()),
-                _ => None,
-            };
-            if let Some(arr) = arr_opt {
-                for elem in &arr.elems {
-                    if let Some(el) = elem {
-                        args.push(swc_ecma_ast::ExprOrSpread {
-                            spread: el.spread,
-                            expr: el.expr.clone(),
-                        });
-                    } else {
-                        // Hole — vira literal 0 (matching JS undefined → 0 here).
-                        args.push(swc_ecma_ast::ExprOrSpread {
-                            spread: None,
-                            expr: Box::new(Expr::Lit(Lit::Num(swc_ecma_ast::Number {
-                                span: Default::default(),
-                                value: 0.0,
-                                raw: Some("0".into()),
-                            }))),
-                        });
-                    }
-                }
-                continue;
-            }
-            // Spread de não-literal/nao-resolvivel: deixa como esta.
-            // Codegen vai rejeitar com erro claro em runtime.
-            args.push(arg);
-        } else {
-            args.push(arg);
-        }
-    }
 }
 
 impl LiftAcc {
