+++ crates/rts-codegen/src/codegen/lower/passes/args/mod.rs
@@ -1,0 +1,6 @@
+//! Passes que reescrevem callsites pra alinhar args com assinatura
+//! do callee: defaults, rest variadicos e spread literal.
+
+pub mod default_args;
+pub mod rest_args;
+pub mod spread_args;
