+++ crates/rts-codegen/src/codegen/lower/passes/mod.rs
@@ -1,0 +1,8 @@
+//! Passes de transformacao do AST antes do codegen.
+//!
+//! Cada submodulo expoe um `pub(crate) fn <name>(program: &mut Program)`
+//! aplicavel ao programa inteiro. Ordem de aplicacao eh definida em
+//! `lower::func::compile_program` — alterar com cuidado, alguns passes
+//! dependem de transformacoes anteriores.
+
+pub mod args;
